fixed some minor unhandled errors, changed scoreboard and player list to remove index numbers, and renamed/moved files for structure. used already created player type in topthree
