+++ frontend/modules/PlayerRoom.tsx
@@ -13,7 +13,7 @@ import { SubmitGuess } from "@/types/dto";
 import { Room } from "@/types";
 import { RoomStatusDisplay } from "./PlayerRoom/RoomStatusDisplay";
 import { RoomStatus } from "@/common/enum";
-import Scoreboard from "./ScoreBoard";
+import TopThreePlayers from "./PlayerRoom/TopThreePlayers";
 
 export default function PlayerRoom({ params }: { params: { roomId: string } }) {
   const router = useRouter();
@@ -107,8 +107,6 @@ export default function PlayerRoom({ params }: { params: { roomId: string } }) {
     (eachPlayer) => eachPlayer.name === playerName
   );
 
-  /**TODO: Need to reset word guess input on submit */
-
   return (
     <div className="min-h-screen p-4 md:p-6 honeycomb-bg">
       <div className="max-w-xl mx-auto">
@@ -181,13 +179,8 @@ export default function PlayerRoom({ params }: { params: { roomId: string } }) {
               <h2 className="card-title">Scoreboard</h2>
               <div className="space-y-4">
                 {room?.status === RoomStatus.ENDED && (
-                  <div className="card-body" style={{padding:'0px'}}>
-                    {/* <div className="w-full flex justify-center">
-                      <h2 className="text-2xl font-sans animate-party">
-                        🎉 Top Three Winners 🏆
-                      </h2>
-                    </div> */}
-                    <Scoreboard winners={room.players} />
+                  <div className="card-body" style={{ padding: "0px" }}>
+                    <TopThreePlayers winners={room.players} />
                   </div>
                 )}
 
@@ -211,7 +204,7 @@ export default function PlayerRoom({ params }: { params: { roomId: string } }) {
                         }`}
                       >
                         <span className="font-medium">
-                          {index + 1}. {player.name}
+                          {player.name}
                           {player.name === playerName && (
                             <span className="badge badge-sm ml-2">You</span>
                           )}
