+++ frontend/modules/PlayerRoom/Confetti.tsx
@@ -1,7 +1,7 @@
-import confetti from 'canvas-confetti';
-import { useEffect } from 'react';
+import confetti from "canvas-confetti";
+import { useEffect } from "react";
 
-const PartySprinkles = () => {
+const Confetti = () => {
   useEffect(() => {
     const duration = 5 * 1000;
     const end = Date.now() + duration;
@@ -13,12 +13,14 @@ const PartySprinkles = () => {
         angle: 90,
         spread: 180,
         origin: { y: 0 },
-        colors: ['#bb0000', '#ffffff', '#00ff00', '#0000ff', '#ffff00'],
+        colors: ["#bb0000", "#ffffff", "#00ff00", "#0000ff", "#ffff00"],
       });
     }, 200);
 
     return () => clearInterval(interval);
   }, []);
+
+  return <></>;
 };
 
-export default PartySprinkles;
+export default Confetti;
