+++ frontend/modules/PlayerRoom/RoomStatusDisplay.tsx
@@ -1,6 +1,6 @@
 import { RoomStatus } from "@/common/enum";
 import React from "react";
-import Confetti from "../Confetti"
+import Confetti from "./Confetti";
 
 /**
  * @file the client side Room Status display component
@@ -28,18 +28,18 @@ export const RoomStatusDisplay = ({
 
   return (
     <>
-    <div className="text-center space-y-4">
-      <div className="flex items-center justify-center">
-        <div
-          className={`w-12 h-12 ${
-            roomStatus === RoomStatus.ENDED ? "bg-green-500" : "bg-primary"
-          } rounded-full transition-colors animate-pulse`}
-        >
-          {roomStatus === RoomStatus.ENDED && <Confetti/>}
+      <div className="text-center space-y-4">
+        <div className="flex items-center justify-center">
+          <div
+            className={`w-12 h-12 ${
+              roomStatus === RoomStatus.ENDED ? "bg-green-500" : "bg-primary"
+            } rounded-full transition-colors animate-pulse`}
+          >
+            {roomStatus === RoomStatus.ENDED && <Confetti />}
           </div>
+        </div>
+        <p className="text-base-content/70 text-xl">{displayText}</p>
       </div>
-      <p className="text-base-content/70 text-xl">{displayText}</p>
-    </div>
     </>
   );
 };
